+++ dict.h
@@ -9,7 +9,7 @@ using namespace std;
 class trie_node{
 public:
     trie_node* par;
-    int word_count;
+    long long word_count;
     vector<trie_node*> child=vector<trie_node*>(54,nullptr);
     trie_node();
     ~trie_node();
@@ -27,7 +27,7 @@ public:
     trie_node* root;
     trie();
     ~trie();
-    int search(string& word);
+    long long search(string& word);
     
 };
 
@@ -48,7 +48,7 @@ public:
 
     void insert_sentence(int book_code, int page, int paragraph, int sentence_no, string sentence);
 
-    int get_word_count(string word);
+    long long get_word_count(string word);
 
     void dump_dictionary(string filename);
